InputStateを分離
